refactored public restaurants route, controllers, and utils
